main: implement EntropySource trait for Onion Messenger
ChaCha20Rng is used because LN already relies on ChaCha20. Ideally
we'd be able to import an EntropySource from LDK, but it's currently
wrapped in the KeysManager struct, which is overkill for what we need.

As with other LDK interfaces that have immutable &self references, a
RefCell is used to provide interior mutability for implementations that
need to make mutable calls.