Add crev note and security audit GH Action for PRs
Encourage the use of `cargo-crev` for auditing dependencies. Crev uses
a web of trust of entities contributing reviews of crates within the
Rust ecosystem in this case. Read more at https://github.com/crev-dev/cargo-crev
and see similar use for `rust-bitcoin` at https://github.com/rust-bitcoin/rust-bitcoin/pull/1098.

This commit also adds a GitHub Action for PRs which runs `cargo-audit` to check
dependencies for any known vulnerabilities along with their severity
and steps to remedy. This should be less noisy than dependabot PRs but we
can maybe consider that if we feel it's also helpful to keep track of
dependency updates.

Work is part of #38.