main: fail if LND does not support onion messages
With the merge of LND/7568, we can now expect nodes running LNDK to
have onion messages advertised. Even if we _could_ set the feature bit
in our announcement, we need the feature to also be included in the
init message sent to peers, so we simplify the code and just error if
it isn't announced.