offers: track active payments instead of offers
Instead of tracking the offers we're paying, we move to instead track active
payments by PaymentId through their lifecycle.

This makes it easier for us to track whether we've already received an invoice
for the payment or not. If we have, we ignore the second invoice so we don't
try to pay it again.

As a byproduct, this change also allows us to pay the same offer multiple at
once, which we didn't allow for before.