Added data_dir configuration options.
In certain deployment scenarios, we may not want (or be able) to write
to the default ~/.lndk data directory. For example, the home directory
may be mounted as read-only, or mounted using ephemeral storage.

This change adds a new `data_dir` configuration option to specify where
LNDK should store data. If this option isn't specified, then falls back
to the default ~/.lndk.