Update `cargo-dist` to `dist` 0.27.1
The `cargo-dist` tool has been renamed to `dist` to reflect the fact
that it is becoming a general purpose tool to distribute binaries.

We update `cargo-dist` to `dist` automatically by running `dist init`.
Note the generation of the new custom configuration file which will
be used by `dist` going forward.