handler: retry `pay_invoice` on failure for each blinded path
Before this commit we used the first blinded path offered in the
invoice to send the payment. If it failed because network dynamics
we would fail the payment for the end user.

With this commit if we fail to send the payment through the route
we will try with a new one until we don't have options to do the
payment.