onion_messenger: place subscribe to custom message in a loop.
When the connection with LND is broken, the subscription to custom
message stream is also broken. Therefore, when the reconnection is done,
it would not be possible to receive custom messages anymore. Placing
the subscription to custom messages into a loop we can guarantee that
after the reconnection to LND, the subscription to the custom message can be
reestablish.